+++ arize/experimental/datasets/__init__.py
@@ -1,0 +1,3 @@
+from .core.client import ArizeDatasetsClient
+
+__all__ = ["ArizeDatasetsClient"]